chore: re-organize files
