+++ frontend/src/store/modules/index.ts
@@ -5,11 +5,6 @@ export * from "./notification";
 export * from "./slowQueryPolicy";
 export * from "./router";
 export * from "./sqlEditor";
-export * from "./sqlEditorV2";
-export * from "./sqlEditorTree";
-export * from "./sqlEditorTab";
-export * from "./sqlEditorQueryHistory";
-export * from "./tab";
 export * from "./uistate";
 export * from "./sqlReview";
 export * from "./onboardingState";
