+++ frontend/src/store/modules/sqlEditor/index.ts
@@ -1,0 +1,5 @@
+export * from "./editor";
+export * from "./tab";
+export * from "./tree";
+export * from "./queryHistory";
+export * from "./webTerminal";
